created dynamic.py
