+++ src/atomate2/openmm/flows/__init__.py
@@ -1,3 +1,4 @@
 """Core Flows for OpenMM module."""
 
-from atomate2.openmm.flows.core import DynamicOpenMMFlowMaker, OpenMMFlowMaker
+from atomate2.openmm.flows.core import OpenMMFlowMaker
+from atomate2.openmm.flows.dynamic import DynamicOpenMMFlowMaker
